fix: rename CardRound5 to CardRound6 and update data fetching function for Round 6
